+++ src/reportdialog.cpp
@@ -197,7 +197,7 @@ void ReportDialog::slotSaveAs() {
   if(!u.isEmpty() && u.isValid()) {
     KConfigGroup config(KGlobal::config(), "ExportOptions");
     bool encode = config.readBoolEntry("EncodeUTF8", true);
-    int oldOpt = m_exporter->options();
+    long oldOpt = m_exporter->options();
 
     // turn utf8 off
     long options = oldOpt & ~Export::ExportUTF8;
